requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
google-generativeai==0.4.0
APScheduler==3.10.4
httpx==0.27.0